Build (some) pages virtually
see #3.

Also this shuffles around the code a bit. `code` is not a good name for
a package imho.
