+++ echemdb/__init__.py
@@ -1,0 +1,1 @@
+from .macros import define_env
